:memo: :: add javadocs in API modules
