--- Midnight-Api/src/main/java/me/iqpizza6349/midnight/constraints/bot/BotStatus.java
@@ -1,11 +1,0 @@
-package me.iqpizza6349.midnight.constraints.bot;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum BotStatus {
-    READY,
-    SHUTDOWN
-}
