--- Midnight-Api/src/main/java/me/iqpizza6349/midnight/event/bot/BotEvent.java
@@ -1,10 +1,0 @@
-package me.iqpizza6349.midnight.event.bot;
-
-import me.iqpizza6349.midnight.constraints.bot.BotStatus;
-import me.iqpizza6349.midnight.event.GeneralEvent;
-
-public interface BotEvent extends GeneralEvent {
-
-    BotStatus getStatus();
-
-}
