+++ Midnight-Api/src/main/java/me/iqpizza6349/midnight/event/chat/ChatEvent.java
@@ -2,5 +2,10 @@ package me.iqpizza6349.midnight.event.chat;
 
 import me.iqpizza6349.midnight.event.GeneralEvent;
 
+/**
+ * This interface is Marking interface that <strong>"Auditing Chat Events"</strong>
+ * @author iqpizza6349
+ * @since 1.0
+ */
 public interface ChatEvent extends GeneralEvent {
 }
