+++ Midnight-Api/src/main/java/me/iqpizza6349/midnight/event/listener/EventListener.java
@@ -3,7 +3,22 @@ package me.iqpizza6349.midnight.event.listener;
 import me.iqpizza6349.midnight.event.GeneralEvent;
 
 /**
- * 봇 이벤트 발생 시, 동작할 행위들을 명시해둔 인터페이스
+ * Handler interface that defines the behavior of receiving and processing
+ * events in the Midnight Framework <br>
+ * {@link me.iqpizza6349.midnight.event.handler.EventHandler} allows you
+ * to handle events asynchronously and conveniently.
+ * <pre>
+ *     {@code EventHandler.addListener(EventListener);}
+ * </pre>
+ *
+ * If it is annoying to set up one by one, it is recommended to inherit
+ * the next abstract class. {@link AbstractEventListener}
+ * The Constructor is automatically registered in EventHandler as soon as
+ * it is called.
+ *
+ * @author iqpizza6349
+ * @since 1.0
+ * @see AbstractEventListener
  */
 public interface EventListener {
