+++ Midnight-Api/src/main/java/me/iqpizza6349/midnight/model/channel/Channel.java
@@ -2,12 +2,29 @@ package me.iqpizza6349.midnight.model.channel;
 
 import me.iqpizza6349.midnight.model.message.Message;
 
+/**
+ * Interface for defining multiple channels
+ * @author iqpizza6349
+ * @since 1.0
+ */
 public interface Channel {
 
+    /**
+     * channel's topic that connect to
+     * @return channel's topic
+     */
     String getTopic();
 
+    /**
+     * send message to topic
+     * @param text message to send
+     */
     void sendMessage(String text);
 
+    /**
+     * send message to topic in framework client (not a normal way)
+     * @param message message to send
+     */
     void sendMessage(Message message);
 
 }
