+++ Midnight-Api/src/main/java/me/iqpizza6349/midnight/model/member/Member.java
@@ -4,6 +4,9 @@ import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.ToString;
 
+/**
+ * Member(Client) class with has only name(sender)
+ */
 @Getter
 @ToString
 @RequiredArgsConstructor
