+++ Midnight-Core/src/main/java/me/iqpizza6349/midnight/core/event/lifecycle/MidnightConstructEventPublisher.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import me.iqpizza6349.midnight.core.configuration.MidnightConfiguration;
 import me.iqpizza6349.midnight.event.GeneralEvent;
-import me.iqpizza6349.midnight.event.bot.BotReadyEvent;
+import me.iqpizza6349.midnight.event.midnight.MidnightReadyEvent;
 import me.iqpizza6349.midnight.event.handler.EventHandler;
 import org.springframework.stereotype.Component;
 
@@ -18,7 +18,7 @@ public class MidnightConstructEventPublisher {
 
     @PostConstruct
     private void midnightApplicationConstruct() {
-        GeneralEvent event = new BotReadyEvent(200, configuration.auditing());
+        GeneralEvent event = new MidnightReadyEvent(200, configuration.auditing());
         EventHandler.callEvent(MidnightConstructEventPublisher.class, event, false);
     }
 }
