+++ Midnight-Core/src/main/java/me/iqpizza6349/midnight/core/event/listener/bot/DefaultBotReadyListener.java
@@ -1,7 +1,7 @@
 package me.iqpizza6349.midnight.core.event.listener.bot;
 
 import me.iqpizza6349.midnight.event.GeneralEvent;
-import me.iqpizza6349.midnight.event.bot.BotReadyEvent;
+import me.iqpizza6349.midnight.event.midnight.MidnightReadyEvent;
 import me.iqpizza6349.midnight.event.listener.AbstractEventListener;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -12,7 +12,7 @@ public class DefaultBotReadyListener extends AbstractEventListener {
 
     @Override
     public void onEvent(GeneralEvent event) {
-        if (event instanceof BotReadyEvent) {
+        if (event instanceof MidnightReadyEvent) {
             LOGGER.info("<{}> is ready!", event.getAuditing().getBotName());
         }
     }
